Specify minimum versions in requirements.txt.
We used to use requirements.txt for frozen dependency testing in Travis-CI.
Given that pecan is a framework, it probably makes more sense to specify
"minimum" versions.  This change extracts the list of minimum dependencies into
a requirements file that can be used in other places, like documentation and
installation tools.

Change-Id: Ie2fef9fd2e5ed261b4cc82139882b226ba787680